Updates to curation-guidlines
Some edits as discussed with the curation steering committee on March 2, 2020.

Note:  Appendix A aligns with the section "Canonical locations for the declared license for package definitions."  

The Maven locations previously referenced have changed, so I deleted those references.  We can add more detail to this section when we update where ClearlyDefined is linking to for maven packages.